models/segmentation/requirements.txt
pdf2image==1.17.0
opencv-python==4.10.0.84
matplotlib==3.7.5
scikit-image==0.21.0
ipython==8.12.3
tensorflow==2.13.1
fastapi==0.95.2
# fastapi==0.112.0
pillow>=8.2.0
python-magic==0.4.15
python-multipart==0.0.9
uvicorn==0.30.5
minio==7.2.7\